# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bookmarks - Social Ebook Reader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or get a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User created or retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithFriendStatus"
  /users/{userId}/friends:
    get:
      operationId: getFriends
      tags: [users]
      summary: Get friends of a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friends list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: addFriend
      tags: [users]
      summary: Add a friend
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendInput"
      responses:
        "200":
          description: Friend added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [users]
      summary: Remove a friend
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: friendId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Friend removed
  /users/{userId}/saved-comments:
    get:
      operationId: getSavedComments
      tags: [users]
      summary: Get saved comments for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
components:
  schemas:
...
    Comment:
      type: object
      required: [id, quoteId, userId, username, avatarColor, text, likeCount, likedByMe, savedByMe, createdAt]
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarColor:
          type: string
        text:
          type: string
        quoteText:
          type: string
        likeCount:
          type: integer
        likedByMe:
          type: boolean
        savedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    User:
      type: object
      required: [id, username, avatarColor, createdAt, token]
      properties:
        id:
          type: integer
        username:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string
          format: date-time
        token:
          type: string
          description: HMAC-signed bearer token — send as Authorization header
    UserWithFriendStatus:
      type: object
      required: [id, username, avatarColor, createdAt, isFriend]
      properties:
        id:
          type: integer
        username:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string
          format: date-time
        isFriend:
          type: boolean
    UserInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
          minLength: 2
    FriendInput:
      type: object
      required: [friendId]
      properties:
        friendId:
          type: integer