# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bookmarks - Social Ebook Reader API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  avatarColor: string;
  createdAt: Date;
  /** HMAC-signed bearer token — send as Authorization header */
  token: string;
}